requirements.txt
openai-agents>=0.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
schedule>=1.2.0
feedparser>=6.0.10
pandas>=2.0.0
openpyxl>=3.1.0


